move levelCounts handle method to MetaDataHandler
