+++ core/src/main/java/io/ddf/content/IHandleSchema.java
@@ -47,12 +47,6 @@ public interface IHandleSchema extends IHandleDDFFunctionalGroup {
 
   Map<String, Map<String, Integer>> computeLevelCounts(String[] columnNames) throws DDFException;
 
-  void removeLevelCounts() throws DDFException;
-
-  void setLevelCounts(Map<String, Map<String, Integer>> levelCounts) throws DDFException;
-
-  Map<String, Map<String, Integer>> getLevelCounts() throws DDFException;
-
   List<Object> computeFactorLevels(String columnName) throws DDFException;
 
   Factor<?> setAsFactor(String columnName) throws DDFException;
