+++ core/src/main/java/io/ddf/content/SchemaHandler.java
@@ -32,8 +32,6 @@ public class SchemaHandler extends ADDFFunctionalGroupHandler implements
 
   private Schema mSchema;
 
-  private Map<String, Map<String, Integer>> mLevelCounts = new HashMap<String, Map<String, Integer>>();
-
   @Override
   public Schema getSchema() {
     return mSchema;
@@ -137,22 +135,6 @@ public class SchemaHandler extends ADDFFunctionalGroupHandler implements
     throw new UnsupportedOperationException();
   }
 
-  @Override
-  public void setLevelCounts(Map<String, Map<String, Integer>> levelCounts) throws DDFException {
-    Map<String, Map<String, Integer>> currentLevelCounts = this.getLevelCounts();
-    currentLevelCounts.putAll(levelCounts);
-  }
-
-  @Override
-  public void removeLevelCounts() throws DDFException {
-    this.mLevelCounts = new HashMap<String, Map<String, Integer>>();
-  }
-
-  @Override
-  public Map<String, Map<String, Integer>> getLevelCounts() throws DDFException {
-    return this.mLevelCounts;
-  }
-
   public List<Object> computeFactorLevels(String columnName) throws DDFException {
     throw new UnsupportedOperationException();
   }
@@ -169,8 +151,12 @@ public class SchemaHandler extends ADDFFunctionalGroupHandler implements
 
   @Override
   public synchronized void unsetAsFactor(int columnIndex) {
-    if (this.getSchema() != null)
-      this.getSchema().getColumn(columnIndex).unsetAsFactor();
+
+    if (this.getSchema() != null) {
+      Column column = this.getSchema().getColumn(columnIndex);
+      column.unsetAsFactor();
+      this.getDDF().getMetaDataHandler().removeLevelCountsForColumn(column.getName());
+    }
   }
 
   @Override
